Reorganize commands folder
