+++ src/commands/index.js
@@ -1,5 +1,3 @@
-export * from './init.command'
-export * from './modules-deploy.command'
-export * from './modules-list.command'
-export * from './scripts-list.command'
-export * from './scripts-run.command'
+export * from './init'
+export * from './modules'
+export * from './scripts'
