+++ src/commands/scripts/index.js
@@ -1,0 +1,2 @@
+export * from './list.command'
+export * from './run.command'
